Use int64_t as fluid_long_long_t in C99

+++ include/fluidsynth/types.h
@@ -73,7 +73,8 @@ typedef __int64 fluid_long_long_t; // even on 32bit windows
  * A typedef for C99's type long long, which is at least 64-bit wide, as guaranteed by the C99.
  * @p __int64 will be used as replacement for VisualStudio 2010 and older.
  */
-typedef long long fluid_long_long_t;
+#include <stdint.h>
+typedef int64_t fluid_long_long_t;
 #endif
 
 /* @} */
